ml/requirements.txt
tensorflow>=2.15.0,<2.18.0
scikit-learn>=1.3.0,<1.6.0
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0
pyyaml>=6.0,<7.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0,<11.0.0
pytest>=7.4.0,<9.0.0
